Updated Latch SDK to its latest version (currently in 'development' branch). It was necessary to comment out a test statement to make it work properly.
