+++ latch_plugin/lib/pairingservice.php
@@ -24,87 +24,86 @@
  * process (it depends on the Latch SDK).
  */
 
-namespace OCA\Latch_Plugin\Lib{
+namespace OCA\Latch_Plugin\Lib;
 
-    use \OCP\IL10N;
-    
-    use \Latch;
+use \OCP\IL10N;
 
-    // Class definition:
-    class PairingService{
+use \ElevenPaths\Latch\LatchApp as Latch;
 
-        private $l10n;
+// Class definition:
+class PairingService{
 
-        private $dbService;
+    private $l10n;
 
-        public function __construct(IL10N $l10n, DbService $dbService) {
-            $this->l10n      = $l10n;
-            $this->dbService = $dbService;
-        }
+    private $dbService;
 
-        public function getLatchToken(){
-            $token = '';
-            if(isset($_POST['latch_token'])){
-                $token = $_POST['latch_token'];
-            }
-
-            return $token;
-        }
-
-        public function pairAccount($token, $user){
-            $msg = ''; // This is the default value of the returned variable
-            $accountID = $this->dbService->retrieveAccountID($user);
-            if(empty($accountID)){ 
-                // Current user account not paired
-                $appID = $this->dbService->retrieveAppID();
-                $appSecret = $this->dbService->retrieveAppSecret();
-
-                if(!empty($appID) && !empty($appSecret)){
-                    // Latch plugin properly configured
-                    $api = new Latch($appID, $appSecret);
-                    $pairResponse = $api->pair($token);
-                    $pairResponseData = $pairResponse->getData();
-                    $msg = $this->processPairResponseData($pairResponseData,$user);
-                }
-            }
-
-            return $msg;
-        }
-
-        private function processPairResponseData($pairResponseData,$user){
-            $msg = ''; // This is the default value of the returned variable
-            if(!empty($pairResponseData) && !empty($pairResponseData->{'accountId'})){
-                // An accountID has been received
-                $accountID = $pairResponseData->{'accountId'};
-                // Save accountID in database and set success message:
-                $this->dbService->saveAccountID($user, $accountID);
-                $msg = ['class' => 'msg success', 'value' => $this->l10n->t('Pairing success')];
-            }else{
-                // There has been no success in the pairing process:
-                $msg = ['class' => 'msg error', 'value' => $this->l10n->t('Pairing token not valid')];
-            }
-
-            return $msg;
-        }
-
-        public function unpairAccount($user){
-            // First, check if user has actually a paired Latch account:
-            $accountID = $this->dbService->retrieveAccountID($user);
-            if(!empty($accountID)){
-                // The current user has a paired account. 
-                // Let's proceed with the unpairing:
-                $appID = $this->dbService->retrieveAppID();
-                $appSecret = $this->dbService->retrieveAppSecret();
-
-                if(!empty($appID) && !empty($appSecret)){
-                    // Latch plugin properly configured
-                    $api = new Latch($appID, $appSecret);
-                    $api->unpair($accountID);
-                    $this->dbService->deleteAccountData($user);
-                }
-            }
-        }
-
-
+    public function __construct(IL10N $l10n, DbService $dbService) {
+        $this->l10n      = $l10n;
+        $this->dbService = $dbService;
     }
+
+    public function getLatchToken(){
+        $token = '';
+        if(isset($_POST['latch_token'])){
+            $token = $_POST['latch_token'];
+        }
+
+        return $token;
+    }
+
+    public function pairAccount($token, $user){
+        $msg = ''; // This is the default value of the returned variable
+        $accountID = $this->dbService->retrieveAccountID($user);
+        if(empty($accountID)){ 
+            // Current user account not paired
+            $appID = $this->dbService->retrieveAppID();
+            $appSecret = $this->dbService->retrieveAppSecret();
+
+            if(!empty($appID) && !empty($appSecret)){
+                // Latch plugin properly configured
+                $api = new Latch($appID, $appSecret);
+                $pairResponse = $api->pair($token);
+                $pairResponseData = $pairResponse->getData();
+                $msg = $this->processPairResponseData($pairResponseData,$user);
+            }
+        }
+
+        return $msg;
+    }
+
+    private function processPairResponseData($pairResponseData,$user){
+        $msg = ''; // This is the default value of the returned variable
+        if(!empty($pairResponseData) && !empty($pairResponseData->{'accountId'})){
+            // An accountID has been received
+            $accountID = $pairResponseData->{'accountId'};
+            // Save accountID in database and set success message:
+            $this->dbService->saveAccountID($user, $accountID);
+            $msg = ['class' => 'msg success', 'value' => $this->l10n->t('Pairing success')];
+        }else{
+            // There has been no success in the pairing process:
+            $msg = ['class' => 'msg error', 'value' => $this->l10n->t('Pairing token not valid')];
+        }
+
+        return $msg;
+    }
+
+    public function unpairAccount($user){
+        // First, check if user has actually a paired Latch account:
+        $accountID = $this->dbService->retrieveAccountID($user);
+        if(!empty($accountID)){
+            // The current user has a paired account. 
+            // Let's proceed with the unpairing:
+            $appID = $this->dbService->retrieveAppID();
+            $appSecret = $this->dbService->retrieveAppSecret();
+
+            if(!empty($appID) && !empty($appSecret)){
+                // Latch plugin properly configured
+                $api = new Latch($appID, $appSecret);
+                $api->unpair($accountID);
+                $this->dbService->deleteAccountData($user);
+            }
+        }
+    }
+
+
 }
